Rename IPointerResult to IInteractorResult and fix namespace
