+++ Runtime/Input/Interactors/IInteractor.cs
@@ -78,7 +78,7 @@ namespace RealityToolkit.Input.Interactors
         RayStep[] Rays { get; }
 
         /// <summary>
-        /// The Physics Layers, in prioritized order, that are used to determine the <see cref="IPointerResult.CurrentPointerTarget"/> when raycasting.
+        /// The Physics Layers, in prioritized order, that are used to determine the <see cref="IInteractorResult.CurrentPointerTarget"/> when raycasting.
         /// </summary>
         /// <remarks>
         /// If set, will override the <see cref="InputServiceProfile.PointerRaycastLayerMasks"/>'s default raycasting layer mask array.
@@ -101,23 +101,23 @@ namespace RealityToolkit.Input.Interactors
         /// <summary>
         /// The combined physics and graphics raycast pointer result.
         /// </summary>
-        IPointerResult Result { get; set; }
+        IInteractorResult Result { get; set; }
 
         /// <summary>
-        /// Is the focus currently locked to the current <see cref="IPointerResult.CurrentPointerTarget"/>?
+        /// Is the focus currently locked to the current <see cref="IInteractorResult.CurrentPointerTarget"/>?
         /// </summary>
         /// <remarks>
-        /// This effectively means that the focused <see cref="IPointerResult.CurrentPointerTarget"/> result cannot be changed,
-        /// even though the focus data may return hit information that is not the <see cref="IPointerResult.CurrentPointerTarget"/>.
+        /// This effectively means that the focused <see cref="IInteractorResult.CurrentPointerTarget"/> result cannot be changed,
+        /// even though the focus data may return hit information that is not the <see cref="IInteractorResult.CurrentPointerTarget"/>.
         /// </remarks>
         bool IsFocusLocked { get; set; }
 
         /// <summary>
-        /// Is the <see cref="IPointerResult.CurrentPointerTarget"/>'s position is locked in sync with the <see cref="Result"/> position?
+        /// Is the <see cref="IInteractorResult.CurrentPointerTarget"/>'s position is locked in sync with the <see cref="Result"/> position?
         /// </summary>
         /// <remarks>
-        /// When true, the <see cref="IPointerResult.CurrentPointerTarget"/>s position will also be updated to move in sync with the <see cref="Result"/> position.
-        /// This position is calculated using the offset of the offset between the <see cref="IPointerResult.CurrentPointerTarget"/> position and the <see cref="Result"/> position.
+        /// When true, the <see cref="IInteractorResult.CurrentPointerTarget"/>s position will also be updated to move in sync with the <see cref="Result"/> position.
+        /// This position is calculated using the offset of the offset between the <see cref="IInteractorResult.CurrentPointerTarget"/> position and the <see cref="Result"/> position.
         /// </remarks>
         GameObject SyncedTarget { get; set; }
